Much work, much progress
The game basically works now, as a stand alone. Pieces can be dragged and drop.  They lock into place, and the move is registered with the back-end scrip. It updates a game string behind the scenes, and checks for wins and for tie games. See issue for outstanding issues. Even the slash across wining rows of pieces in working now.